requirements.txt
paho-mqtt
Flask